LPS-157939 Move the test to the proper module
We are testing the export-import of layouts it is not about the layout page templates, so we are going to move the test to the proper module
